lesson-20/file-upload-and-download

+++ controllers/admin.js
@@ -1,4 +1,6 @@
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
+
 const { validationResult } = require('express-validator');
 const Product = require('../models/product');
 
@@ -6,15 +8,35 @@ exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
     pageTitle: 'Add Product',
     path: '/admin/add-product',
-    editing: false
+    editing: false,
+    hasError: false,
+    errorMessage: null,
+    validationErrors: []
   });
 };
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file; //binary data
   const price = req.body.price;
   const description = req.body.description;
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description
+      },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: []
+    });
+  }
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -25,14 +47,16 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title,
-        imageUrl,
+        image,
         price,
         description
       },
-      errorMessage: 'Database operation failed, please try again.',
+      errorMessage: errors.array()[0].msg,
       validationErrors: []
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     // _id: new mongoose.Types.ObjectId('67913d3dc6880fcc37ce870f'),
@@ -74,7 +98,9 @@ exports.getEditProduct = (req, res, next) => {
         path: '/admin/edit-product',
         editing: editMode,
         hasError: false,
-        product
+        product,
+        errorMessage: null,
+        validationErrors: []
       });
     })
     .catch((err) => {
@@ -87,23 +113,23 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const productId = req.body.productId;
   const updatedTitle = req.body.title;
-  const updatedImageUrl = req.body.imageUrl;
   const updatedPrice = req.body.price;
+  const image = req.file;
   const updatedDescription = req.body.description;
 
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
-      pageTitle: 'Add Product',
+      pageTitle: 'Edit Product',
       path: '/admin/edit-product',
       editing: false,
       hasError: true,
       product: {
-        title,
-        imageUrl,
-        price,
-        description
+        title: updatedTitle,
+        price: updatedPrice,
+        description: updatedDescription,
+        _id: productId
       },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array()
@@ -118,9 +144,14 @@ exports.postEditProduct = (req, res, next) => {
       }
 
       product.title = updatedTitle;
-      product.imageUrl = updatedImageUrl;
       product.price = updatedPrice;
       product.description = updatedDescription;
+
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl); // delete old image
+        product.imageUrl = image.path;
+      }
+
       return product.save().then(() => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
@@ -155,7 +186,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
+
+  Product.findById(productId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl); // delete image
+
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
